requirements.txt
PyOpenGL
PyOpenGL-accelerate
glfw
pyassimp
numpy